requirements.txt
# Django core
Django==4.2.11
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-cors-headers==4.3.1
django-storages==1.14.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Celery and Redis
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Selenium automation
selenium==4.17.2
webdriver-manager==4.0.1

# File handling
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.2.0

# Email
django-anymail==10.2

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-factoryboy==2.6.0
factory-boy==3.3.0
faker==22.5.0
responses==0.24.1

# API docs
drf-spectacular==0.27.1

# Security
django-axes==6.3.0
django-otp==1.3.0

# Utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
boto3==1.34.25

# Data export
openpyxl==3.1.2
